Reduce anchor search service ownership
Split search and n-hop anchor responsibilities into focused composition helpers while preserving existing service import paths and behavior.

Constraint: Issue #630 requires anchor-search-service.ts and n-hop-search-service.ts to stay under 500 lines with public import/export compatibility preserved.

Rejected: Public API reshaping | would broaden the migration surface beyond the requested composition split.

Confidence: high

Scope-risk: moderate

Directive: Keep anchor orchestration files as compatibility surfaces; place future search/reanchor or linked-memory logic in focused helper modules.

Tested: npm test -- packages/memento-core/src/domains/anchor/services/anchor/anchor-search-service.spec.ts packages/memento-core/src/domains/anchor/services/anchor/n-hop-search-service.spec.ts; npm test -- packages/memento-core/src/test/architecture/dependency-boundaries.spec.ts; npm run lint; npm run type-check; npm test; graphify rebuild; graphify anchor connection distribution check; git diff --check; npm run check-debt-markers -- --production-only

Not-tested: No GitHub push or PR creation was performed.
